refactor: migrate `commonTags` to v2
Common tags are moved to a dedicated file in the shared folder.
