+++ src/v2/components/password/index.ts
@@ -1,7 +1,7 @@
 import * as aws from '@pulumi/aws-v7';
 import * as pulumi from '@pulumi/pulumi';
 import * as random from '@pulumi/random';
-import { commonTags } from '../../../constants';
+import { commonTags } from '../../shared/common-tags';
 
 export namespace Password {
   export type Args = {
