+++ src/v2/components/redis/elasticache-redis.ts
@@ -1,7 +1,7 @@
 import * as aws from '@pulumi/aws-v7';
 import * as pulumi from '@pulumi/pulumi';
 import * as awsx from '@pulumi/awsx-v3';
-import { commonTags } from '../../../constants';
+import { commonTags } from '../../shared/common-tags';
 import { mergeWithDefaults } from '../../shared/merge-with-defaults';
 
 export namespace ElastiCacheRedis {
