+++ src/v2/components/static-site/s3-assets.ts
@@ -1,6 +1,6 @@
 import * as pulumi from '@pulumi/pulumi';
 import * as aws from '@pulumi/aws-v7';
-import { commonTags } from '../../../constants';
+import { commonTags } from '../../shared/common-tags';
 
 export namespace S3Assets {
   export type Args = {
