+++ src/v2/components/vpc/index.ts
@@ -1,6 +1,6 @@
 import * as pulumi from '@pulumi/pulumi';
 import * as awsx from '@pulumi/awsx-v3';
-import { commonTags } from '../../../constants';
+import { commonTags } from '../../shared/common-tags';
 import { enums } from '@pulumi/awsx-v3/types';
 import { mergeWithDefaults } from '../../shared/merge-with-defaults';
