+++ src/v2/shared/common-tags.ts
@@ -1,0 +1,6 @@
+import * as pulumi from '@pulumi/pulumi';
+
+export const commonTags = {
+  Env: pulumi.getStack(),
+  Project: pulumi.getProject(),
+};
